backend/pyproject.toml
[project]
name = "bank-chat"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.15",
    "fastapi>=0.116.1",
    "google-adk>=1.8.0",
    "google-generativeai>=0.8.5",
    "psycopg2-binary==2.9.10",
    "uvicorn>=0.35.0",
]
